Curve Fitting: sync with upstream
Correct accidental float use

+++ extern/curve_fit_nd/intern/curve_fit_cubic.c
@@ -623,13 +623,13 @@ static void cubic_from_points_offset_fallback(
 		}
 	}
 
-	float alpha_l = (dists[0] / 0.75) /  dot_vnvn(tan_l, a[0], dims);
-	float alpha_r = (dists[1] / 0.75) / -dot_vnvn(tan_r, a[1], dims);
+	double alpha_l = (dists[0] / 0.75) /  dot_vnvn(tan_l, a[0], dims);
+	double alpha_r = (dists[1] / 0.75) / -dot_vnvn(tan_r, a[1], dims);
 
-	if (!(alpha_l > 0.0f)) {
+	if (!(alpha_l > 0.0)) {
 		alpha_l = dir_dist / 3.0;
 	}
-	if (!(alpha_r > 0.0f)) {
+	if (!(alpha_r > 0.0)) {
 		alpha_r = dir_dist / 3.0;
 	}
